requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==24.1.1
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.1
distlib==0.3.8
fastapi==0.109.2
filelock==3.13.1
flake8==7.0.0
h11==0.14.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy==1.8.0
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.9.13
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.0
pycodestyle==2.11.1
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
virtualenv==20.25.0
